Deduplicate favorite card feature editors
