+++ src/panels/lovelace/editor/config-elements/hui-cover-tilt-favorite-card-feature-editor.ts
@@ -1,41 +1,12 @@
-import { html, LitElement, nothing } from "lit";
-import { customElement, property, state } from "lit/decorators";
-import "../../../../components/ha-alert";
-import type { HomeAssistant } from "../../../../types";
-import type {
-  CoverTiltFavoriteCardFeatureConfig,
-  LovelaceCardFeatureContext,
-} from "../../card-features/types";
-import type { LovelaceCardFeatureEditor } from "../../types";
+import { customElement } from "lit/decorators";
+import type { LocalizeKeys } from "../../../../common/translations/localize";
+import type { CoverTiltFavoriteCardFeatureConfig } from "../../card-features/types";
+import { HuiFavoriteCardFeatureEditorBase } from "./hui-favorite-card-feature-editor-base";
 
 @customElement("hui-cover-tilt-favorite-card-feature-editor")
-export class HuiCoverTiltFavoriteCardFeatureEditor
-  extends LitElement
-  implements LovelaceCardFeatureEditor
-{
-  @property({ attribute: false }) public hass?: HomeAssistant;
-
-  @property({ attribute: false }) public context?: LovelaceCardFeatureContext;
-
-  @state() private _config?: CoverTiltFavoriteCardFeatureConfig;
-
-  public setConfig(config: CoverTiltFavoriteCardFeatureConfig): void {
-    this._config = config;
-  }
-
-  protected render() {
-    if (!this.hass || !this._config) {
-      return nothing;
-    }
-
-    return html`
-      <ha-alert alert-type="info">
-        ${this.hass.localize(
-          "ui.panel.lovelace.editor.features.types.cover-tilt-favorite.description"
-        )}
-      </ha-alert>
-    `;
-  }
+export class HuiCoverTiltFavoriteCardFeatureEditor extends HuiFavoriteCardFeatureEditorBase<CoverTiltFavoriteCardFeatureConfig> {
+  protected readonly _descriptionKey =
+    "ui.panel.lovelace.editor.features.types.cover-tilt-favorite.description" satisfies LocalizeKeys;
 }
 
 declare global {
