+++ src/panels/lovelace/editor/config-elements/hui-valve-position-favorite-card-feature-editor.ts
@@ -1,41 +1,12 @@
-import { html, LitElement, nothing } from "lit";
-import { customElement, property, state } from "lit/decorators";
-import "../../../../components/ha-alert";
-import type { HomeAssistant } from "../../../../types";
-import type {
-  LovelaceCardFeatureContext,
-  ValvePositionFavoriteCardFeatureConfig,
-} from "../../card-features/types";
-import type { LovelaceCardFeatureEditor } from "../../types";
+import { customElement } from "lit/decorators";
+import type { LocalizeKeys } from "../../../../common/translations/localize";
+import type { ValvePositionFavoriteCardFeatureConfig } from "../../card-features/types";
+import { HuiFavoriteCardFeatureEditorBase } from "./hui-favorite-card-feature-editor-base";
 
 @customElement("hui-valve-position-favorite-card-feature-editor")
-export class HuiValvePositionFavoriteCardFeatureEditor
-  extends LitElement
-  implements LovelaceCardFeatureEditor
-{
-  @property({ attribute: false }) public hass?: HomeAssistant;
-
-  @property({ attribute: false }) public context?: LovelaceCardFeatureContext;
-
-  @state() private _config?: ValvePositionFavoriteCardFeatureConfig;
-
-  public setConfig(config: ValvePositionFavoriteCardFeatureConfig): void {
-    this._config = config;
-  }
-
-  protected render() {
-    if (!this.hass || !this._config) {
-      return nothing;
-    }
-
-    return html`
-      <ha-alert alert-type="info">
-        ${this.hass.localize(
-          "ui.panel.lovelace.editor.features.types.valve-position-favorite.description"
-        )}
-      </ha-alert>
-    `;
-  }
+export class HuiValvePositionFavoriteCardFeatureEditor extends HuiFavoriteCardFeatureEditorBase<ValvePositionFavoriteCardFeatureConfig> {
+  protected readonly _descriptionKey =
+    "ui.panel.lovelace.editor.features.types.valve-position-favorite.description" satisfies LocalizeKeys;
 }
 
 declare global {
